most recent version of code

+++ index.js
@@ -1,9 +1,7 @@
 #! /usr/bin/env node
 
 'use strict';
-// this is for development only
-const perf = require('execution-time')();
-perf.start();
+
 
 const fs = require('fs-extra');
 const chalk = require('chalk');
@@ -20,23 +18,37 @@ const touch = require('touch');
 
 program
   .version(version)
-  .usage('new component_name [options]')
+  .usage('help for commands and options')
   .command('new <name>')
   .option('-p, --pure', 'creates a pure component')
   .description('generates a new component')
-  .action(function (name) {
+  .action(function (name, cmd) {
     const options = config.get();
 
+    if (cmd.pure)
+      options.componentType = 'pure';
     buildReactComponent(name, options);
   });
 
 program
-  .option('-c, --config', 'Set up configuration', setConfigOptions)
-  .option('--print', 'Print your current config options', printConfigOptions);
+  .command('config')
+  .description('allows you to set up your own personalize configuration')
+  .action(function () {
+    setConfigOptions();
+  });
+
+
+program
+  .command('print')
+  .description('Prints out your currently configured options')
+  .action(function () {
+    printConfigOptions();
+  });
+
 
 
 program.arguments('<arg>').action((arg) => {
-  console.log(chalk.red(`\nInvalid command "${arg}", please below for valid commands.\n`));
+  // console.log(chalk.red(`\nInvalid command "${arg}", please below for valid commands.\n`));
   program.outputHelp();
 });
 
@@ -50,6 +62,9 @@ if (!program.args.length) {
 
 
 
+
+
+
 // Builds the react component and it's styling
 function buildReactComponent(name, options) {
   const capitalizedName = name.replace(/^\w/, (c) => c.toUpperCase());
@@ -57,17 +72,26 @@ function buildReactComponent(name, options) {
   const reactFileName = `${currentPath}/${capitalizedName}${options.jsExtensions}`;
   const cssFileName = `${currentPath}/${capitalizedName}${options.cssType}`;
 
-  fs.outputFile(reactFileName, componetFile)
-    .then(() => {
-      console.log(chalk.blue(`Succesfully created ${reactFileName}!`));
-      if (options.cssType !== 'none') {
-        touch(cssFileName, '');
-        console.log(chalk.blue(`Succesfully created ${cssFileName}`));
-      }
-    })
-    .catch((err) => {
-      throw err;
-    });
+
+  if (fs.existsSync(reactFileName)) {
+    console.log(chalk.red(`\nERROR: React file already exists at: "${reactFileName}"\n`));
+  } else {
+    fs.outputFile(reactFileName, componetFile)
+      .then(() => {
+        console.log(chalk.green(`\nSUCCESS: Created ${capitalizedName}${options.jsExtensions}`));
+
+        if (fs.existsSync(cssFileName)) {
+          console.log(chalk.red(`ERROR: CSS file already exists at "${cssFileName}"`));
+        } else if (options.cssType !== 'none') {
+          touch(cssFileName, '');
+          console.log(chalk.green(`SUCCESS: Created ${capitalizedName}${options.cssType}`));
+        }
+        console.log();
+      })
+      .catch((err) => {
+        throw err;
+      });
+  }
 }
 
 
@@ -85,6 +109,3 @@ function printConfigOptions() {
   console.log(config.get());
   console.log();
 }
-
-const results = perf.stop();
-console.log(results.time);
